+++ app/src/main/java/com/sopt/now/data/service/AuthService.kt
@@ -1,5 +1,7 @@
-package com.sopt.now.data
+package com.sopt.now.data.service
 
+import com.sopt.now.data.model.request.SignUpRequestDto
+import com.sopt.now.data.model.response.SignUpResponseDto
 import retrofit2.Call
 import retrofit2.http.Body
 import retrofit2.http.POST
