+++ app/src/main/java/com/sopt/now/presentation/home/FriendDiffCallback.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.adapter
+package com.sopt.now.presentation.home
 
 import androidx.recyclerview.widget.DiffUtil
 import com.sopt.now.data.FriendInfoData
